Moves all main utilities to a new file
- this fixes an import error in utilities
- imports in utilities and __main__ are updated accordingly
